+++ unix/Xvnc/programs/Xserver/hw/vnc/init.c
@@ -5,6 +5,7 @@
  */
 
 /* Copyright (C) 2009-2022 D. R. Commander.  All Rights Reserved.
+ * Copyright (C) 2021 Steffen Kieß
  * Copyright (C) 2016-2017 Pierre Ossman for Cendio AB.  All Rights Reserved.
  * Copyright (C) 2010 University Corporation for Atmospheric Research.
  *                    All Rights Reserved.
@@ -357,6 +358,22 @@ int ddxProcessArgument(int argc, char *argv[], int i)
   if (strcasecmp(argv[i], "-rfbport") == 0) {  /* -rfbport port */
     REQUIRE_ARG();
     rfbPort = atoi(argv[i + 1]);
+    return 2;
+  }
+
+  if (strcasecmp(argv[i], "-rfbunixpath") == 0) {  /* -rfbunixpath path */
+    REQUIRE_ARG();
+    rfbUDSPath = argv[i + 1];
+    return 2;
+  }
+
+  if (strcasecmp(argv[i], "-rfbunixmode") == 0) {  /* -rfbunixmode mode */
+    char *ptr;
+    REQUIRE_ARG();
+    ptr = argv[i + 1];
+    rfbUDSMode = strtol(argv[i + 1], &ptr, 8);
+    if (*argv[i + 1] == 0 || *ptr != 0 || rfbUDSMode < 0 || rfbUDSMode > 0777)
+      FatalError("Invalid mode %s\n", argv[i + 1]);
     return 2;
   }
 
@@ -1578,6 +1595,8 @@ void ddxGiveUp(enum ExitCode error)
     sprintf(unixSocketName, "/tmp/.X11-unix/X%s", display);
     unlink(unixSocketName);
   }
+  if (rfbUDSPath && rfbUDSCreated)
+    unlink(rfbUDSPath);
 }
 
 
@@ -1669,7 +1688,9 @@ void ddxUseMsg(void)
   ErrorF("                       selection (typically used when pasting with the middle\n");
   ErrorF("                       mouse button)\n");
   ErrorF("-noreverse             disable reverse connections\n");
-  ErrorF("-rfbport port          TCP port for RFB protocol\n");
+  ErrorF("-rfbport port          TCP port for RFB connections\n");
+  ErrorF("-rfbunixpath path      path to Unix domain socket for RFB connections\n");
+  ErrorF("-rfbunixmode mode      Unix domain socket permissions\n");
   ErrorF("-rfbwait time          max time in ms to wait for a send/receive operation\n");
   ErrorF("                       to/from a connected viewer to complete [default: %d]\n",
          DEFAULT_MAX_CLIENT_WAIT);