+++ unix/Xvnc/programs/Xserver/hw/vnc/rfb.h
@@ -1014,6 +1014,9 @@ extern int rfbMaxClientConnections;
 extern int rfbMaxClientWait;
 
 extern int rfbPort;
+extern const char *rfbUDSPath;
+extern int rfbUDSMode;
+extern Bool rfbUDSCreated;
 extern int rfbListenSock;
 
 extern void rfbInitSockets(void);
@@ -1028,7 +1031,6 @@ extern int ReadExact(rfbClientPtr cl, char *buf, int len);
 extern int ReadExactTimeout(rfbClientPtr cl, char *buf, int len, int timeout);
 extern int SkipExact(rfbClientPtr cl, int len);
 extern int WriteExact(rfbClientPtr cl, char *buf, int len);
-extern int ListenOnTCPPort(int port);
 extern int ConnectToTcpAddr(char *host, int port);
 
 extern const char *sockaddr_string(rfbSockAddr *addr, char *buf, int len);